Começando a componentização e Atualizando Layout
