+++ src/Components/Banner/index.js
@@ -1,0 +1,11 @@
+import './styles.css'
+export default function Banner(){
+    return(
+        <div className="banner">
+            <div className="foto"/>
+            <div className="head-line">
+                <h1>AQUI VOCÊ ENCONTRA TUDO SOBRE SEUS HERÓIS E VILÕES FAVORITOS!</h1>
+            </div>
+        </div>
+    )
+}
