+++ src/Components/Banner/styles.css
@@ -1,0 +1,41 @@
+@import url('https://fonts.googleapis.com/css2?family=Carter+One&display=swap');
+.banner{
+    width: 100%;
+    height: 85vh;
+    position: relative;
+    overflow: hidden;
+    margin-top: 15vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    z-index: 0;
+}
+.banner .foto{
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    background-image: url('../../Assets/thumb.png');
+    background-size: cover;
+    background-position: initial;
+    background-repeat: no-repeat;
+    filter: sepia(0.1) contrast(0.8) hue-rotate(3deg);
+}
+.banner .head-line{
+    width: 100%;
+    height: 100%;
+    position: relative;
+    display: flex;
+    flex-direction: column;   
+    justify-content: center; 
+    align-items: center;
+}
+.banner h1{
+    max-width: 60%;
+    margin-bottom: 1rem;
+    font-size: 3em;
+    font-weight: 800;
+    font-family: 'Carter One', cursive;
+    text-align: center;
+    text-shadow: -8px 6px 0px black;
+    color: white;
+}
