+++ src/Components/Cards/styles.css
@@ -1,0 +1,26 @@
+.heroes-list{
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(200px, 240px));
+    grid-gap: 3.5rem;
+    text-align: center;
+    padding: 0;
+    margin-top: 3rem;
+    justify-content: center;
+}
+.heroes-list li{
+    min-height: 300px;
+    max-height: 360px;
+    border: 1px solid rgb(210,210,210);
+    border-radius: 15px;
+    box-shadow: 0 0 15px 0 rgb(150,150,150);
+    overflow: hidden;
+}
+.heroes-list img{
+    width: 100%;
+}
+.heroes-list h3{
+    font-size: 1.5em;
+    padding: 0 0 10px 0;
+    margin: 0;
+    color: rgb(100,100,100)
+}
