+++ src/index.css
@@ -1,8 +1,15 @@
-html, body {
+*{
+  -webkit-transition: 0.2s ease-in-out;
+  -moz-transition: 0.2s ease-in-out;
+  -o-transition: 0.2s ease-in-out;
+  transition: 0.2s ease-in-out;
+  list-style-type: none;
+}
+html, body, ul, li , h1, h2, h3, a{
   margin: 0;
   padding: 0;
   border: 0;
-  font-size: 100%;  
+  text-decoration: none;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
